Added PCC cluster to the helper.js in order to get IM callbacks

+++ examples/pump-controller-app/cc13x2x7_26x2x7/main/ZclCallbacks.cpp
@@ -31,7 +31,7 @@
 using namespace ::chip;
 
 void emberAfPostAttributeChangeCallback(EndpointId endpoint, ClusterId clusterId, AttributeId attributeId, uint8_t mask,
-                                        uint16_t manufacturerCode, uint8_t type, uint8_t size, uint8_t * value)
+                                        uint16_t manufacturerCode, uint8_t type, uint16_t size, uint8_t * value)
 {
     if (clusterId != ZCL_ON_OFF_CLUSTER_ID)
     {
